Move relevant data functions to module
